+++ ListOfCompanyEmployees/AddOrEditDepartment.xaml
@@ -1,9 +1,8 @@
-﻿<Window x:Class="ListOfCompanyEmployees.AddDepartmentForm"
+﻿<Window x:Class="ListOfCompanyEmployees.AddOrEditDepartment"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:ListOfCompanyEmployees"
         mc:Ignorable="d"
         Title="Добавьте (измените) департамент!!!" Height="400" Width="600">
