+++ ListOfCompanyEmployees/App.xaml
@@ -1,7 +1,7 @@
 ﻿<Application x:Class="ListOfCompanyEmployees.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="Window1.xaml"
+             StartupUri="MainWindow.xaml"
              >
     <Application.Resources>
         <ResourceDictionary>
