+++ ListOfCompanyEmployees/MainWindow.xaml
@@ -1,27 +1,52 @@
 ﻿<Window x:Class="ListOfCompanyEmployees.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:ListOfCompanyEmployees"
-        mc:Ignorable="d"
-        Title="Список сотрудников компании" Height="400" Width="600">
-    <Grid>
+        Height="800"
+        Width="800"
+        Title="Ресурсы"
+        Background="{DynamicResource Brush}">
+
+    <Grid Grid.Row="1">
         <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="1*"/>            
-            <ColumnDefinition Width="1*"/>
+            <ColumnDefinition Width="*"/>
+            <ColumnDefinition Width="*"/>
         </Grid.ColumnDefinitions>
-        <ListBox Name="Department" VerticalAlignment="Top" Margin="0,25,0,0" Height="204" SelectionChanged="Department_SelectionChanged">
-            
-        </ListBox>
-        <ListBox Name="Employee" VerticalAlignment="Top" Grid.Column="1" Margin="0,25,0,0" Height="204">
-             
-        </ListBox>
-        <Label Content="Департамент" HorizontalAlignment="Left" Margin="114,0,0,0" VerticalAlignment="Top"/>
-        <Label Content="Сотрудники" Grid.Column="1" HorizontalAlignment="Left" Margin="108,0,0,0" VerticalAlignment="Top"/>
-        <Button Content="Добавить департамент" HorizontalAlignment="Left" Margin="50,318,0,0" VerticalAlignment="Top" Width="200" Click="AddDepartment">
-            
-        </Button>  
+
+        <StackPanel Grid.Column="0">
+            <ListView Margin="10" Name="lbEmployee" 
+                     VerticalAlignment="Top" 
+                     Height="250">
+                <ListView.View>
+                    <GridView>
+                        <GridViewColumn Header="№" DisplayMemberBinding="{Binding Id}"/>
+                        <GridViewColumn Header="Имя" Width="85" DisplayMemberBinding="{Binding Name}"/>
+                        <GridViewColumn Header="Возраст" Width="85" DisplayMemberBinding="{Binding Age}"/>
+                        <GridViewColumn Header="Департамент" Width="85"  DisplayMemberBinding="{Binding Department}"/>
+                        <GridViewColumn Header="Зарплата" Width="85" DisplayMemberBinding="{Binding Salary}"/>                        
+                    </GridView>
+                </ListView.View>
+            </ListView>
+            <Button Margin="10" Height="25" Width="70" 
+                    VerticalAlignment="Bottom"
+                    Click="AddNewEmployee">Добавить</Button>
+            <Button Content="Изменить" Height="25" Width="70" Click="ChangeEmployee"/>
+        </StackPanel>
+
+        <StackPanel Grid.Column="1">
+            <ListView Margin="10" x:Name="lbDepartmant" 
+                     VerticalAlignment="Top" 
+                     Height="250">
+                <ListView.View>
+                    <GridView>
+                        <GridViewColumn Header="Имя" Width="350" DisplayMemberBinding="{Binding Name}"/>
+                    </GridView>
+                </ListView.View>
+            </ListView>
+            <Button Margin="10" Height="25" Width="70" 
+                    VerticalAlignment="Bottom"
+                    Click="AddNewDepartments" Content="Добавить"/>
+            <Button Content="Изменить" Height="25" Width="70" Click="ChangeDepartment"/>
+        </StackPanel>
 
     </Grid>
