--- ListOfCompanyEmployees/Window1.xaml
@@ -1,54 +1,0 @@
-﻿<Window x:Class="ListOfCompanyEmployees.Window1"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:sys="clr-namespace:System;assembly=mscorlib"
-        Height="800"
-        Width="800"
-        Title="Ресурсы"
-        Background="{DynamicResource Brush}">
-
-    <Grid Grid.Row="1">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="*"/>
-            <ColumnDefinition Width="*"/>
-        </Grid.ColumnDefinitions>
-
-        <StackPanel Grid.Column="0">
-            <ListView Margin="10" Name="lbEmployee" 
-                     VerticalAlignment="Top" 
-                     Height="250">
-                <ListView.View>
-                    <GridView>
-                        <GridViewColumn Header="№" DisplayMemberBinding="{Binding Id}"/>
-                        <GridViewColumn Header="Имя" Width="85" DisplayMemberBinding="{Binding Name}"/>
-                        <GridViewColumn Header="Возраст" Width="85" DisplayMemberBinding="{Binding Age}"/>
-                        <GridViewColumn Header="Департамент" Width="85"  DisplayMemberBinding="{Binding Department}"/>
-                        <GridViewColumn Header="Зарплата" Width="85" DisplayMemberBinding="{Binding Salary}"/>                        
-                    </GridView>
-                </ListView.View>
-            </ListView>
-            <Button Margin="10" Height="25" Width="70" 
-                    VerticalAlignment="Bottom"
-                    Click="AddNewEmployee">Добавить</Button>
-            <Button Content="Изменить" Height="25" Width="70" Click="ChangeEmployee"/>
-        </StackPanel>
-
-        <StackPanel Grid.Column="1">
-            <ListView Margin="10" x:Name="lbDepartmant" 
-                     VerticalAlignment="Top" 
-                     Height="250">
-                <ListView.View>
-                    <GridView>
-                        <GridViewColumn Header="Имя" Width="350" DisplayMemberBinding="{Binding Name}"/>
-                    </GridView>
-                </ListView.View>
-            </ListView>
-            <Button Margin="10" Height="25" Width="70" 
-                    VerticalAlignment="Bottom"
-                    Click="AddNewDepartments" Content="Добавить"/>
-            <Button Content="Изменить" Height="25" Width="70" Click="ChangeDepartment"/>
-        </StackPanel>
-
-    </Grid>
-
-</Window>
